+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Latest - ???
 ### Added
+- SC2227: Warn about redirections in the middle of 'find' commands
 - SC2224,SC2225,SC2226: Warn when using mv/cp/ln without a destination
 - SC2223: Quote warning specific to `: ${var=value}`
 - SC1127: Warn about C-style comments